added catologue related entities
