--- apps/task-nest/src/auth/entities/auth.entity.ts
@@ -1,1 +1,0 @@
-export class Auth {}
