+++ apps/task-nest/src/auth/entities/session.entity.ts
@@ -1,0 +1,5 @@
+import { Entity } from 'typeorm';
+import { AbstractEntity } from '@app/common';
+
+@Entity()
+export class SessionEntity extends AbstractEntity {}
